Extract GraphQL alias tests into their own file
