+++ packages/apollo-link-state/src/__tests__/index.ts
@@ -15,14 +15,6 @@ const query = gql`
   }
 `;
 
-const aliasedQuery = gql`
-  query Test {
-    fie: foo @client {
-      bar
-    }
-  }
-`;
-
 const doubleQuery = gql`
   query Double {
     foo @client {
@@ -180,53 +172,6 @@ it('runs resolvers for missing client queries with variables', done => {
   });
   execute(client, { query, variables: { id: 1 } }).subscribe(({ data }) => {
     expect(data).toEqual({ foo: { bar: 1 } });
-    done();
-  }, done.fail);
-});
-
-it('runs resolvers for missing client queries with aliased field', done => {
-  const query = gql`
-    query Aliased {
-      foo @client {
-        bar
-      }
-      baz: bar {
-        foo
-      }
-    }
-  `;
-  const sample = new ApolloLink(() =>
-    //The server takes care of the aliasing, so returns baz, not bar
-    Observable.of({ data: { baz: { foo: true } } }),
-  );
-  const client = withClientState({
-    resolvers,
-  });
-  execute(client.concat(sample), { query }).subscribe(({ data }) => {
-    try {
-      expect(data).toEqual({ foo: { bar: true }, baz: { foo: true } });
-    } catch (e) {
-      done.fail(e);
-    }
-    done();
-  }, done.fail);
-});
-
-it('runs resolvers for client queries when aliases are in use on the @client-tagged node', done => {
-  const client = withClientState({
-    resolvers: {
-      Query: {
-        foo: () => ({ bar: true }),
-        fie: () => {
-          done.fail(
-            "Called the resolver using the alias' name, instead of the correct resolver name.",
-          );
-        },
-      },
-    },
-  });
-  execute(client, { query: aliasedQuery }).subscribe(({ data }) => {
-    expect(data).toEqual({ fie: { bar: true } });
     done();
   }, done.fail);
 });
